Create universal routes object in flask
