+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+addopts = -sv
+filterwarnings =
+	ignore::DeprecationWarning:flask_wtf.*:
