+++ src/utils/strings/config_strs.py
@@ -1,0 +1,8 @@
+# Config Env vars labels
+class CONFIG_ENVS:
+    DATABASE_URL_DEV = "DATABASE_URL_DEV"
+    DATABASE_URL_TEST = "DATABASE_URL_TEST"
+    SECRET_KEY = "SECRET_KEY"
+    MAILJET_SECRET_KEY = "MAILJET_SECRET_KEY"
+    MAILJET_API_KEY = "MAILJET_API_KEY"
+    BASE_EMAIL = "BASE_EMAIL"
