+++ src/utils/strings/html_identifiers.py
@@ -1,0 +1,5 @@
+
+# General Identifiers
+class IDENTIFIERS:
+    SPLASH_PAGE = "Welcome to URLS4IRL"
+    HTML_404 = "404 - Invalid Request"
