+++ src/utils/strings/model_strs.py
@@ -1,0 +1,45 @@
+
+# Strings for standardizing the model serialization
+ID = "id"
+NAME = "name"
+CREATED_BY = "created_by"
+CREATED_AT = "created_at"
+DESCRIPTION = "description"
+EMAIL = "email"
+MEMBERS = "members"
+PASSWORD = "password"
+TAG_STRING = "tag_string"
+TAGGED_URL = "tagged_url"
+TAG = "Tag"
+TAGS = "tags"
+URLS = "urls"
+URL = "URL"
+URL_TAGS = "url_tags"
+URL_STRING = "url_string"
+URL_ID = "url_ID"
+URL_TITLE = "url_title"
+ADDED_BY = "Added_by"
+USERNAME = "username"
+UTUB_DESCRIPTION = "UTub_description"
+
+
+class MODELS:
+    ID = ID
+    NAME = NAME
+    URL_ID = URL_ID
+    URL_TAGS = URL_TAGS
+    URL_STRING = URL_STRING
+    ADDED_BY = ADDED_BY
+    URL_TITLE = URL_TITLE
+    TAGGED_URL = TAGGED_URL
+    USERNAME = USERNAME
+    TAG = TAG
+    CREATED_AT = CREATED_AT
+    CREATED_BY = CREATED_BY
+    DESCRIPTION = DESCRIPTION
+    MEMBERS = MEMBERS
+    URLS = URLS
+    TAGS = TAGS
+    URL = URL
+    TAG_STRING = TAG_STRING
+    UTUB_DESCRIPTION = UTUB_DESCRIPTION
