+++ tests/integration/test_add_url_to_utub_route.py
@@ -3,14 +3,11 @@ from flask_login import current_user
 
 from src.models import Utub, URLS, Utub_Urls
 from tests.models_for_test import valid_url_strings
-from src.utils import strings as U4I_STRINGS
-
-URL_FORM = U4I_STRINGS.URL_FORM
-URL_SUCCESS = U4I_STRINGS.URL_SUCCESS
-STD_JSON = U4I_STRINGS.STD_JSON_RESPONSE
-MODEL_STRS = U4I_STRINGS.MODELS
-URL_FAILURE = U4I_STRINGS.URL_FAILURE
-ADD_URL_URL = "urls.add_url"
+from src.utils.all_routes import ROUTES
+from src.utils.strings.form_strs import URL_FORM
+from src.utils.strings.json_strs import STD_JSON_RESPONSE as STD_JSON
+from src.utils.strings.model_strs import MODELS as MODEL_STRS
+from src.utils.strings.url_strs import URL_FAILURE, URL_SUCCESS
 
 
 def test_add_valid_url_as_utub_member(
@@ -90,7 +87,7 @@ def test_add_valid_url_as_utub_member(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 200
@@ -209,7 +206,7 @@ def test_add_valid_url_as_utub_creator(
         URL_FORM.URL_TITLE: url_title_to_add,
     }
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 200
@@ -318,7 +315,7 @@ def test_add_invalid_url_as_utub_member(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 400
@@ -417,7 +414,7 @@ def test_add_invalid_url_as_utub_creator(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 400
@@ -488,7 +485,7 @@ def test_add_valid_url_to_nonexistent_utub(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 404
@@ -553,7 +550,7 @@ def test_add_valid_url_to_utub_not_a_member_of(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=id_of_utub_not_member_of), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=id_of_utub_not_member_of), data=add_url_form
     )
 
     assert add_url_response.status_code == 403
@@ -653,7 +650,7 @@ def test_add_fresh_url_to_utub(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=id_of_utub_that_is_creator_of),
+        url_for(ROUTES.URLS.ADD_URL, utub_id=id_of_utub_that_is_creator_of),
         data=add_url_form,
     )
 
@@ -754,7 +751,7 @@ def test_add_duplicate_url_to_utub_as_same_user_who_added_url(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=id_of_utub_that_is_creator_of),
+        url_for(ROUTES.URLS.ADD_URL, utub_id=id_of_utub_that_is_creator_of),
         data=add_url_form,
     )
 
@@ -844,7 +841,7 @@ def test_add_duplicate_url_to_utub_as_creator_of_utub_not_url_adder(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=id_of_utub_that_is_creator_of),
+        url_for(ROUTES.URLS.ADD_URL, utub_id=id_of_utub_that_is_creator_of),
         data=add_url_form,
     )
 
@@ -934,7 +931,7 @@ def test_add_duplicate_url_to_utub_as_member_of_utub_not_url_adder(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=id_of_utub_that_is_member_of), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=id_of_utub_that_is_member_of), data=add_url_form
     )
 
     assert add_url_response.status_code == 400
@@ -1024,7 +1021,7 @@ def test_add_url_missing_url(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 400
@@ -1103,7 +1100,7 @@ def test_add_url_missing_url_title(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     assert add_url_response.status_code == 400
@@ -1182,7 +1179,7 @@ def test_add_url_missing_csrf_token(
     }
 
     add_url_response = client.post(
-        url_for(ADD_URL_URL, utub_id=utub_id_to_add_to), data=add_url_form
+        url_for(ROUTES.URLS.ADD_URL, utub_id=utub_id_to_add_to), data=add_url_form
     )
 
     # Assert invalid response code
