+++ tests/integration/test_email_validation.py
@@ -7,21 +7,13 @@ from tests.utils_for_test import get_csrf_token
 from src import db
 from src.models import User
 from src.utils.constants import EMAIL_CONSTANTS
-from src.utils import strings as U4I_STRINGS
+from src.utils.all_routes import ROUTES
+from src.utils.strings.json_strs import STD_JSON_RESPONSE as STD_JSON
+from src.utils.strings.html_identifiers import IDENTIFIERS
+from src.utils.strings.splash_form_strs import REGISTER_FORM
+from src.utils.strings.email_validation_strs import EMAILS, EMAILS_FAILURE
+from src.utils.strings.user_strs import USER_FAILURE
 
-STD_JSON = U4I_STRINGS.STD_JSON_RESPONSE
-REGISTER_FORM = U4I_STRINGS.REGISTER_FORM
-USER_FAILURE = U4I_STRINGS.USER_FAILURE
-EMAILS = U4I_STRINGS.EMAILS
-EMAILS_FAILURE = U4I_STRINGS.EMAILS_FAILURE
-IDENTIFIERS = U4I_STRINGS.IDENTIFIERS
-LOGIN_URL = "splash.login"
-REGISTER_URL = "splash.register_user"
-SPLASH_URL = "splash.splash_page"
-CONFIRM_EMAIL_URL = "splash.confirm_email_after_register"
-SEND_VALIDATION_EMAIL_URL = "splash.send_validation_email"
-VALIDATE_EMAIL_URL = "splash.validate_email"
-HOME_PAGE_URL = "utubs.home"
 VALIDATE_EMAIL_MODAL_TITLE = '<h1 class="modal-title validate-email-text validate-email-title">Validate Your Email!</h1>'
 
 def test_registered_user_is_not_email_validated(app, load_register_page):
@@ -35,7 +27,7 @@ def test_registered_user_is_not_email_validated(app, load_register_page):
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token
 
     response = client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
 
     # Assert user gets shown email validation modal
@@ -59,14 +51,14 @@ def test_registered_not_email_validated_user_access_splash_page(load_register_pa
 
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token_string
     response = client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
 
     # Correctly sends URL to email validation modal
     assert response.status_code == 201
     assert VALIDATE_EMAIL_MODAL_TITLE.encode() in response.data
 
-    response = client.get(url_for(SPLASH_URL))
+    response = client.get(url_for(ROUTES.SPLASH.SPLASH_PAGE))
     assert response.status_code == 200
     assert IDENTIFIERS.SPLASH_PAGE.encode() in response.data
     assert f"{EMAILS.EMAIL_VALIDATION_MODAL_CALL}".encode() in response.data
@@ -82,16 +74,16 @@ def test_registered_not_email_validated_user_access_home_page(load_register_page
 
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token_string
     response = client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
 
     # Correctly sends URL to email validation modal
     assert response.status_code == 201
     assert VALIDATE_EMAIL_MODAL_TITLE.encode() in response.data
 
-    response = client.get(url_for(HOME_PAGE_URL), follow_redirects=True)
+    response = client.get(url_for(ROUTES.UTUBS.HOME), follow_redirects=True)
     assert response.history[0].status_code == 302
-    assert response.history[0].location == url_for(SPLASH_URL)
+    assert response.history[0].location == url_for(ROUTES.SPLASH.SPLASH_PAGE)
     assert response.status_code == 200
     assert IDENTIFIERS.SPLASH_PAGE.encode() in response.data
     assert f"{EMAILS.EMAIL_VALIDATION_MODAL_CALL}".encode() in response.data
@@ -107,20 +99,20 @@ def test_registered_not_email_validated_user_access_register_login(load_register
 
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token_string
     response = client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
 
     # Correctly sends URL to email validation modal
     assert response.status_code == 201
     assert VALIDATE_EMAIL_MODAL_TITLE.encode() in response.data
 
-    urls_to_check = (url_for(REGISTER_URL), url_for(LOGIN_URL))
+    urls_to_check = (url_for(ROUTES.SPLASH.REGISTER), url_for(ROUTES.SPLASH.LOGIN))
 
     for url in urls_to_check:
         response = client.get(url, follow_redirects=True)
         assert response.history[0].status_code == 302
         assert response.history[0].location == url_for(
-           CONFIRM_EMAIL_URL 
+           ROUTES.SPLASH.CONFIRM_EMAIL 
         )
         assert response.status_code == 200
         assert VALIDATE_EMAIL_MODAL_TITLE.encode() in response.data
@@ -145,7 +137,7 @@ def test_registered_not_email_validated_tries_registering_again(
     client, csrf_token_string = load_register_page
 
     registered_user[REGISTER_FORM.CSRF_TOKEN] = csrf_token_string
-    response = client.post(url_for(REGISTER_URL), data=valid_user_1)
+    response = client.post(url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1)
 
     # Ensure json response from server is valid
     register_user_response_json = response.json
@@ -178,7 +170,7 @@ def test_registered_not_email_validated_tries_logging_in(
     client, csrf_token_string = load_login_page
 
     registered_user[REGISTER_FORM.CSRF_TOKEN] = csrf_token_string
-    response = client.post(url_for(LOGIN_URL), data=valid_user_1)
+    response = client.post(url_for(ROUTES.SPLASH.LOGIN), data=valid_user_1)
 
     # Ensure json response from server is valid
     login_user_response_json = response.json
@@ -224,7 +216,7 @@ def test_token_validates_user(app, load_register_page):
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token
 
     client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
 
     with app.app_context():
@@ -235,12 +227,12 @@ def test_token_validates_user(app, load_register_page):
         assert not user.is_email_authenticated() and not user.email_confirm.is_validated
 
     response = client.get(
-        url_for(VALIDATE_EMAIL_URL, token=user_token), follow_redirects=True
+        url_for(ROUTES.SPLASH.VALIDATE_EMAIL, token=user_token), follow_redirects=True
     )
 
     assert len(response.history) == 1
     assert response.history[0].status_code == 302
-    assert response.history[0].location == url_for(HOME_PAGE_URL)
+    assert response.history[0].location == url_for(ROUTES.UTUBS.HOME)
     assert response.status_code == 200
 
     # Ensure user logged in
@@ -292,11 +284,11 @@ def test_success_on_send_of_email(app, load_register_page):
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token
 
     register_response = client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
     new_csrf_token = get_csrf_token(register_response.data)
     send_email_response = client.post(
-        url_for(SEND_VALIDATION_EMAIL_URL),
+        url_for(ROUTES.SPLASH.SEND_VALIDATION_EMAIL),
         data={REGISTER_FORM.CSRF_TOKEN: csrf_token},
     )
 
@@ -324,15 +316,15 @@ def test_min_rate_limiting_of_sending_email(app, load_register_page):
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token
 
     client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
     client.post(
-        url_for(SEND_VALIDATION_EMAIL_URL),
+        url_for(ROUTES.SPLASH.SEND_VALIDATION_EMAIL),
         data={REGISTER_FORM.CSRF_TOKEN: csrf_token},
     )
 
     send_second_email_response = client.post(
-        url_for(SEND_VALIDATION_EMAIL_URL),
+        url_for(ROUTES.SPLASH.SEND_VALIDATION_EMAIL),
         data={REGISTER_FORM.CSRF_TOKEN: csrf_token},
     )
     second_email_send_json = send_second_email_response.json
@@ -365,7 +357,7 @@ def test_max_rate_limiting_of_sending_email(app, load_register_page):
     valid_user_1[REGISTER_FORM.CSRF_TOKEN] = csrf_token
 
     client.post(
-        url_for(REGISTER_URL), data=valid_user_1, follow_redirects=True
+        url_for(ROUTES.SPLASH.REGISTER), data=valid_user_1, follow_redirects=True
     )
 
     with app.app_context():
@@ -377,7 +369,7 @@ def test_max_rate_limiting_of_sending_email(app, load_register_page):
         db.session.commit()
 
     email_response = client.post(
-        url_for(SEND_VALIDATION_EMAIL_URL),
+        url_for(ROUTES.SPLASH.SEND_VALIDATION_EMAIL),
         data={REGISTER_FORM.CSRF_TOKEN: csrf_token},
     )
     email_response_json = email_response.json
