+++ tests/integration/test_forgot_password.py
@@ -5,18 +5,13 @@ from tests.models_for_test import valid_user_1
 from src import db
 from src.models import User, ForgotPassword
 from src.utils import constants as U4I_CONSTANTS
-from src.utils import strings as U4I_STRINGS
+from src.utils.all_routes import ROUTES
+from src.utils.strings.splash_form_strs import REGISTER_FORM
+from src.utils.strings.json_strs import STD_JSON_RESPONSE as STD_JSON
+from src.utils.strings.reset_password_strs import FORGOT_PASSWORD, RESET_PASSWORD
 
-STD_JSON = U4I_STRINGS.STD_JSON_RESPONSE
-REGISTER_FORM = U4I_STRINGS.REGISTER_FORM
-RESET_PASSWORD = U4I_STRINGS.RESET_PASSWORD
-FORGOT_PASSWORD = U4I_STRINGS.FORGOT_PASSWORD
 USER_CONSTANTS = U4I_CONSTANTS.USER_CONSTANTS
 FORGOT_PASSWORD_MODAL_TITLE = '<h4 class="modal-title">Forgot your password?</h4>'
-FORGOT_PASSWORD_URL = "splash.forgot_password"
-HOME_URL = "utubs.home"
-CONFIRM_EMAIL_URL = "splash.confirm_email_after_register"
-
 
 def test_user_logged_in_email_validated_cannot_access_forgot_password(
     register_first_user, login_first_user_without_register
@@ -30,14 +25,14 @@ def test_user_logged_in_email_validated_cannot_access_forgot_password(
     client, _, _, _ = login_first_user_without_register
 
     forgot_password_response = client.get(
-        url_for(FORGOT_PASSWORD_URL), follow_redirects=True
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE), follow_redirects=True
     )
 
     # Only one redirect to user home page
     assert len(forgot_password_response.history) == 1
     redirect_response = forgot_password_response.history[0]
 
-    assert redirect_response.location == url_for(HOME_URL)
+    assert redirect_response.location == url_for(ROUTES.UTUBS.HOME)
     assert redirect_response.status_code == 302
 
 
@@ -67,14 +62,14 @@ def test_user_registered_not_email_validated_cannot_access_forgot_password(
     assert register_response.status_code == 201
 
     forgot_password_response = client.get(
-        url_for(FORGOT_PASSWORD_URL), follow_redirects=True
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE), follow_redirects=True
     )
 
     # Only one redirect to user home page
     assert len(forgot_password_response.history) == 1
     redirect_response = forgot_password_response.history[0]
 
-    assert redirect_response.location == url_for(CONFIRM_EMAIL_URL)
+    assert redirect_response.location == url_for(ROUTES.SPLASH.CONFIRM_EMAIL)
     assert redirect_response.status_code == 302
 
 
@@ -87,7 +82,7 @@ def test_valid_user_requests_forgot_password_form(register_first_user, load_logi
     """
     client, _ = load_login_page
 
-    forgot_password_response = client.get(url_for(FORGOT_PASSWORD_URL))
+    forgot_password_response = client.get(url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE))
 
     assert forgot_password_response.status_code == 200
     assert FORGOT_PASSWORD_MODAL_TITLE.encode() in forgot_password_response.data
@@ -104,10 +99,10 @@ def test_valid_user_posts_forgot_password_form_without_csrf(
     new_user, _ = register_first_user
     client, _ = load_login_page
 
-    client.get(url_for(FORGOT_PASSWORD_URL))
+    client.get(url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE))
 
     forgot_password_post_response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL]},
     )
 
@@ -137,7 +132,7 @@ def test_forgot_password_with_invalid_email_fails(load_login_page):
     improper_email = "Cat"
 
     response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: improper_email,
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -180,7 +175,7 @@ def test_forgot_password_with_email_not_in_database(app, load_login_page):
         num_of_forgot_password_objs = len(ForgotPassword.query.all())
 
     response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: nonregistered_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -226,7 +221,7 @@ def test_forgot_password_with_validated_email(
         user_id = user.id
 
     response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -275,7 +270,7 @@ def test_forgot_password_with_non_validated_email(
         user_id = user.id
 
     response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -324,7 +319,7 @@ def test_forgot_password_rate_limits_correctly(
 
     initial_send_time = datetime.utcnow()
     first_response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -333,7 +328,7 @@ def test_forgot_password_rate_limits_correctly(
 
     second_send_time = datetime.utcnow()
     second_response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -392,7 +387,7 @@ def test_forgot_password_generates_token_correctly(
         user_id = user.id
 
     forgot_password_response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: valid_user_1[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -448,7 +443,7 @@ def test_user_requests_reset_after_password_reset_object_older_than_hour(
     ) = user_attempts_reset_password_one_hour_old
 
     forgot_password_response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
@@ -505,7 +500,7 @@ def test_two_forgot_password_attempts_more_than_minute_apart_increments_attempts
         db.session.commit()
 
     forgot_password_response = client.post(
-        url_for(FORGOT_PASSWORD_URL),
+        url_for(ROUTES.SPLASH.FORGOT_PASSWORD_PAGE),
         data={
             FORGOT_PASSWORD.EMAIL: new_user[FORGOT_PASSWORD.EMAIL],
             FORGOT_PASSWORD.CSRF_TOKEN: csrf_token,
