+++ tests/integration/test_login_user.py
@@ -6,16 +6,11 @@ from pytest import raises
 
 from tests.models_for_test import invalid_user_1, valid_user_1
 from tests.utils_for_test import get_csrf_token
-from src.utils import strings as U4I_STRINGS
 from src.models import User
-
-LOGIN_FORM = U4I_STRINGS.LOGIN_FORM
-STD_JSON = U4I_STRINGS.STD_JSON_RESPONSE
-LOGIN_FAILURE = U4I_STRINGS.USER_FAILURE
-LOGIN_URL = "splash.login"
-REGISTER_URL = "splash.register_user"
-SPLASH_URL = "splash.splash_page"
-HOME_PAGE_URL = "utubs.home"
+from src.utils.all_routes import ROUTES
+from src.utils.strings.json_strs import STD_JSON_RESPONSE as STD_JSON
+from src.utils.strings.splash_form_strs import LOGIN_FORM
+from src.utils.strings.user_strs import USER_FAILURE
 
 
 def test_login_registered_and_logged_in_user(login_first_user_with_register):
@@ -29,12 +24,12 @@ def test_login_registered_and_logged_in_user(login_first_user_with_register):
     new_user = deepcopy(valid_user_1)
     new_user[LOGIN_FORM.CSRF_TOKEN] = csrf_token_str
 
-    response = client.post(url_for(LOGIN_URL), data=new_user, follow_redirects=True)
+    response = client.post(url_for(ROUTES.SPLASH.LOGIN), data=new_user, follow_redirects=True)
 
     # Correctly responds with URL to home page
     assert len(response.history) == 1
     redirect = response.history[-1]
-    assert redirect.location == url_for(HOME_PAGE_URL)
+    assert redirect.location == url_for(ROUTES.UTUBS.HOME)
     assert redirect.status_code == 302
     assert response.status_code == 200
 
@@ -61,7 +56,7 @@ def test_login_unregistered_user(load_login_page):
     Proper JSON response is as follows:
     {
         STD_JSON.STATUS : STD_JSON.FAILURE,
-        STD_JSON.MESSAGE: LOGIN_FAILURE.UNABLE_TO_LOGIN,
+        STD_JSON.MESSAGE: USER_FAILURE.UNABLE_TO_LOGIN,
         STD_JSON.ERROR_CODE: Integer representing the failure code, 2 for invalid form inputs
         STD_JSON.ERRORS: Array containing objects for each field and their specific error. For example:
             [
@@ -77,7 +72,7 @@ def test_login_unregistered_user(load_login_page):
     invalid_user[LOGIN_FORM.CSRF_TOKEN] = csrf_token_str
 
     response = client.post(
-        url_for(LOGIN_URL),
+        url_for(ROUTES.SPLASH.LOGIN),
         data={
             LOGIN_FORM.CSRF_TOKEN: invalid_user[LOGIN_FORM.CSRF_TOKEN],
             LOGIN_FORM.USERNAME: invalid_user[LOGIN_FORM.USERNAME],
@@ -88,10 +83,10 @@ def test_login_unregistered_user(load_login_page):
     # Ensure json response from server is valid
     login_user_response_json = response.json
     assert login_user_response_json[STD_JSON.STATUS] == STD_JSON.FAILURE
-    assert login_user_response_json[STD_JSON.MESSAGE] == LOGIN_FAILURE.UNABLE_TO_LOGIN
+    assert login_user_response_json[STD_JSON.MESSAGE] == USER_FAILURE.UNABLE_TO_LOGIN
     assert int(login_user_response_json[STD_JSON.ERROR_CODE]) == 2
     assert (
-        LOGIN_FAILURE.USER_NOT_EXIST
+        USER_FAILURE.USER_NOT_EXIST
         in login_user_response_json[STD_JSON.ERRORS][LOGIN_FORM.USERNAME]
     )
 
@@ -111,7 +106,7 @@ def test_login_user_wrong_password(register_first_user, load_login_page):
     Proper JSON response is as follows:
     {
         STD_JSON.STATUS : STD_JSON.FAILURE,
-        STD_JSON.MESSAGE: LOGIN_FAILURE.UNABLE_TO_LOGIN,
+        STD_JSON.MESSAGE: USER_FAILURE.UNABLE_TO_LOGIN,
         STD_JSON.ERROR_CODE: Integer representing the failure code, 2 for invalid form inputs
         STD_JSON.ERRORS: Array containing objects for each field and their specific error. For example:
             [
@@ -127,7 +122,7 @@ def test_login_user_wrong_password(register_first_user, load_login_page):
     new_user[LOGIN_FORM.CSRF_TOKEN] = csrf_token_str
 
     response = client.post(
-        url_for(LOGIN_URL),
+        url_for(ROUTES.SPLASH.LOGIN),
         data={
             LOGIN_FORM.CSRF_TOKEN: new_user[LOGIN_FORM.CSRF_TOKEN],
             LOGIN_FORM.USERNAME: new_user[LOGIN_FORM.USERNAME],
@@ -138,10 +133,10 @@ def test_login_user_wrong_password(register_first_user, load_login_page):
     # Ensure json response from server is valid
     login_user_response_json = response.json
     assert login_user_response_json[STD_JSON.STATUS] == STD_JSON.FAILURE
-    assert login_user_response_json[STD_JSON.MESSAGE] == LOGIN_FAILURE.UNABLE_TO_LOGIN
+    assert login_user_response_json[STD_JSON.MESSAGE] == USER_FAILURE.UNABLE_TO_LOGIN
     assert int(login_user_response_json[STD_JSON.ERROR_CODE]) == 2
     assert (
-        LOGIN_FAILURE.INVALID_PASSWORD
+        USER_FAILURE.INVALID_PASSWORD
         in login_user_response_json[STD_JSON.ERRORS][LOGIN_FORM.PASSWORD]
     )
 
@@ -161,7 +156,7 @@ def test_login_user_missing_csrf(register_first_user, load_login_page):
     client, _ = load_login_page
 
     response = client.post(
-        url_for(LOGIN_URL),
+        url_for(ROUTES.SPLASH.LOGIN),
         data={
             LOGIN_FORM.USERNAME: valid_user_1[LOGIN_FORM.USERNAME],
             LOGIN_FORM.PASSWORD: "A",
@@ -187,17 +182,17 @@ def test_already_logged_in_user_to_splash_page(login_first_user_with_register):
     client, _, logged_in_user, _ = login_first_user_with_register
 
     # Ensure redirect on home page access
-    response = client.get(url_for(SPLASH_URL), follow_redirects=True)
+    response = client.get(url_for(ROUTES.SPLASH.SPLASH_PAGE), follow_redirects=True)
 
     # Correctly redirects first to login page
     # Since already logged in, redirects to home page
     assert len(response.history) == 1
     assert response.history[0].status_code == 302
-    assert response.history[0].request.path == url_for(SPLASH_URL)
+    assert response.history[0].request.path == url_for(ROUTES.SPLASH.SPLASH_PAGE)
 
     # Ensure lands on user's home page
     assert response.status_code == 200
-    assert response.request.path == url_for(HOME_PAGE_URL)
+    assert response.request.path == url_for(ROUTES.UTUBS.HOME)
 
     # Test if user logged in
     assert current_user.username == logged_in_user.username
@@ -216,17 +211,17 @@ def test_already_logged_in_user_to_login_page(login_first_user_with_register):
     client, _, logged_in_user, _ = login_first_user_with_register
 
     # Ensure redirect on home page access
-    response = client.get(url_for(LOGIN_URL), follow_redirects=True)
+    response = client.get(url_for(ROUTES.SPLASH.LOGIN), follow_redirects=True)
 
     # Correctly redirects first to login page
     # Since already logged in, redirects to home page
     assert len(response.history) == 1
     assert response.history[0].status_code == 302
-    assert response.history[0].request.path == url_for(LOGIN_URL)
+    assert response.history[0].request.path == url_for(ROUTES.SPLASH.LOGIN)
 
     # Ensure lands on user's home page
     assert response.status_code == 200
-    assert response.request.path == url_for(HOME_PAGE_URL)
+    assert response.request.path == url_for(ROUTES.UTUBS.HOME)
 
     # Test if user logged in
     assert current_user.username == logged_in_user.username
@@ -245,17 +240,17 @@ def test_already_logged_in_user_to_register_page(login_first_user_with_register)
     client, _, logged_in_user, _ = login_first_user_with_register
 
     # Ensure redirect on home page access
-    response = client.get(url_for(REGISTER_URL), follow_redirects=True)
+    response = client.get(url_for(ROUTES.SPLASH.REGISTER), follow_redirects=True)
 
     # Correctly redirects first to login page
     # Since already logged in, redirects to home page
     assert len(response.history) == 1
     assert response.history[0].status_code == 302
-    assert response.history[0].request.path == url_for(REGISTER_URL)
+    assert response.history[0].request.path == url_for(ROUTES.SPLASH.REGISTER)
 
     # Ensure lands on user's home page
     assert response.status_code == 200
-    assert response.request.path == url_for(HOME_PAGE_URL)
+    assert response.request.path == url_for(ROUTES.UTUBS.HOME)
 
     # Test if user logged in
     assert current_user.username == logged_in_user.username
@@ -273,11 +268,11 @@ def test_already_logged_in_user_to_home_page(login_first_user_with_register):
     client, _, logged_in_user, _ = login_first_user_with_register
 
     # Ensure redirect on home page access
-    response = client.get(url_for(HOME_PAGE_URL), follow_redirects=True)
+    response = client.get(url_for(ROUTES.UTUBS.HOME), follow_redirects=True)
 
     assert len(response.history) == 0
     assert response.status_code == 200
-    assert request.path == url_for(HOME_PAGE_URL)
+    assert request.path == url_for(ROUTES.UTUBS.HOME)
 
     # Test if user logged in
     assert current_user.username == logged_in_user.username
@@ -305,7 +300,7 @@ def test_user_can_logout_after_login(login_first_user_with_register):
 
     # Ensure lands on splash page
     assert response.status_code == 200
-    assert response.request.path == url_for(SPLASH_URL)
+    assert response.request.path == url_for(ROUTES.SPLASH.SPLASH_PAGE)
 
     # Test if user logged in
     with raises(AttributeError):
@@ -331,9 +326,9 @@ def test_user_can_login_logout_login(login_first_user_with_register):
 
     # Ensure lands on splash page
     assert response.status_code == 200
-    assert response.request.path == url_for(SPLASH_URL)
+    assert response.request.path == url_for(ROUTES.SPLASH.SPLASH_PAGE)
 
-    response = client.get(url_for(LOGIN_URL))
+    response = client.get(url_for(ROUTES.SPLASH.LOGIN))
 
     # Ensure on login page
     assert (
@@ -348,14 +343,14 @@ def test_user_can_login_logout_login(login_first_user_with_register):
         b'<input class="form-control login-register-form-group" id="password" name="password" required type="password" value="">'
         in response.data
     )
-    assert response.request.path == url_for(LOGIN_URL)
+    assert response.request.path == url_for(ROUTES.SPLASH.LOGIN)
 
     # Grab csrf token from login page
     valid_user_1[LOGIN_FORM.CSRF_TOKEN] = get_csrf_token(response.data)
 
     # Post data to login page
     response = client.post(
-        url_for(LOGIN_URL),
+        url_for(ROUTES.SPLASH.LOGIN),
         data={
             LOGIN_FORM.CSRF_TOKEN: valid_user_1[LOGIN_FORM.CSRF_TOKEN],
             LOGIN_FORM.USERNAME: valid_user_1[LOGIN_FORM.USERNAME],
@@ -366,7 +361,7 @@ def test_user_can_login_logout_login(login_first_user_with_register):
 
     # Ensure backend sends url to home page for frontend to redirect to
     assert response.status_code == 200
-    assert response.data == bytes(f"{url_for(HOME_PAGE_URL)}", "utf-8")
+    assert response.data == bytes(f"{url_for(ROUTES.UTUBS.HOME)}", "utf-8")
 
     # test if user logged in
     assert current_user.username == logged_in_user.username
@@ -382,8 +377,8 @@ def test_login_modal_is_shown(app_with_server_name, client):
     """
     with client:
         with app_with_server_name.app_context():
-            client.get(url_for(SPLASH_URL))
-            response = client.get(url_for(LOGIN_URL))
+            client.get(url_for(ROUTES.SPLASH.SPLASH_PAGE))
+            response = client.get(url_for(ROUTES.SPLASH.LOGIN))
         assert (
             b'<form id="ModalForm" method="POST" class="login-register-form" action="" novalidate>'
             in response.data
@@ -402,7 +397,7 @@ def test_login_modal_is_shown(app_with_server_name, client):
             b'<input class="form-control login-register-form-group" id="password" name="password" required type="password" value="">'
             in response.data
         )
-        assert request.path == url_for(LOGIN_URL)
+        assert request.path == url_for(ROUTES.SPLASH.LOGIN)
 
 
 def test_login_modal_logs_user_in(app_with_server_name, client, register_first_user):
@@ -414,16 +409,16 @@ def test_login_modal_logs_user_in(app_with_server_name, client, register_first_u
     registered_user_data, _ = register_first_user
     with client:
         with app_with_server_name.app_context():
-            client.get(url_for(SPLASH_URL))
-            response = client.get(url_for(LOGIN_URL))
+            client.get(url_for(ROUTES.SPLASH.SPLASH_PAGE))
+            response = client.get(url_for(ROUTES.SPLASH.LOGIN))
         csrf_token = get_csrf_token(response.data)
 
         registered_user_data[LOGIN_FORM.CSRF_TOKEN] = csrf_token
 
-        response = client.post(url_for(LOGIN_URL), data=registered_user_data)
+        response = client.post(url_for(ROUTES.SPLASH.LOGIN), data=registered_user_data)
 
         assert response.status_code == 200
-        assert response.data == bytes(f"{url_for(HOME_PAGE_URL)}", "utf-8")
+        assert response.data == bytes(f"{url_for(ROUTES.UTUBS.HOME)}", "utf-8")
 
         assert current_user.username == registered_user_data[LOGIN_FORM.USERNAME]
         assert check_password_hash(
