+++ tests/unit/test_login_form.py
@@ -1,10 +1,8 @@
 from flask import url_for, request
 
-from src.utils import strings as U4I_STRINGS
-
-LOGIN_FORM = U4I_STRINGS.LOGIN_FORM
-STD_JSON = U4I_STRINGS.STD_JSON_RESPONSE
-LOGIN_FAILURE = U4I_STRINGS.USER_FAILURE
+from src.utils.strings.json_strs import STD_JSON_RESPONSE as STD_JSON
+from src.utils.strings.splash_form_strs import LOGIN_FORM
+from src.utils.strings.user_strs import USER_FAILURE as LOGIN_FAILURE
 
 
 def test_login_no_password(load_login_page):
